Extracted common handler functionality into separate class
